Moved helper methods to subclasses to have codec specific implementation which supports fallback
